Bot/requirements.txt
# Latest discord.py
git+https://github.com/Maverun/Pyanimu
aioredis
redis
PTable
pillow
osuapi
unidecode
GitPython
psutil
lxml
BeautifulSoup4
